src/backend/requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
scikit-learn
transformers
torch
python-dotenv
requests